backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.50.0
watchdog>=4.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
pyyaml>=6.0.2
python-multipart>=0.0.9
rich>=13.8.0
aiosqlite>=0.20.0
websockets>=13.0
python-dotenv>=1.0.0
